refactor: remove notion workspace and migrate to sanity cms
