+++ .env.example
@@ -1,11 +1,2 @@
-NOTION_TOKEN=
-NOTION_DB_ID=
-
-NOTION_TOKEN_PARTNERS=
-NOTION_DB_PARTNERS_ID=
-
-NOTION_TOKEN_RESEARCH=
-NOTION_DB_RESEARCH_ID=
-
 NEXT_PUBLIC_SANITY_PROJECT_ID=
 NEXT_PUBLIC_SANITY_DATASET=
