+++ docs/contributor-guide.md
@@ -8,7 +8,7 @@ Use this document when you want the shortest path from "I need to change X" to t
 - `src/views/` owns page composition
 - `src/components/` owns sections and reusable UI
 - `src/data/` owns static content in Git
-- `packages/notion-data/` owns direct Notion access and field mapping
+- `src/sanity/` owns the CMS schemas and Studio configuration
 - `src/styles/index.css` owns the shared visual system
 
 If you are unsure where to edit, start from the rendered route in `src/app/`, then trace into its view and section components.
@@ -21,10 +21,10 @@ pnpm dev
 pnpm lint
 pnpm test
 pnpm verify
-pnpm typecheck:all
+pnpm typecheck
 ```
 
-Use `pnpm verify` before handoff for the normal repo gate. Use `pnpm typecheck:all` when you changed TypeScript shapes or cross-package contracts.
+Use `pnpm verify` before handoff for the normal repo gate. Use `pnpm typecheck` when you changed TypeScript shapes.
 
 ## Common Changes
 
@@ -34,7 +34,7 @@ Use `pnpm verify` before handoff for the normal repo gate. Use `pnpm typecheck:a
 2. Add or reuse a page composition component in `src/views/`
 3. Build metadata and JSON-LD in `src/config/seo.ts`
 4. Add navigation only if the page should appear in the header or footer
-5. Run `pnpm verify` and `pnpm typecheck:all`
+5. Run `pnpm verify` and `pnpm typecheck`
 
 Pattern to copy:
 
@@ -57,26 +57,25 @@ Examples:
 
 - FAQs: `src/data/qanda.tsx`, `src/data/apply/faq.tsx`, `src/data/e-lab/FAQ.tsx`
 - homepage copy/images: `src/data/homepage.tsx`
-- partner groupings: `src/data/partners.tsx`
 - community copy: `src/data/community.tsx`
 
 If copy is hard-coded inside a section component, move it into `src/data/` only if that makes the code easier to maintain. Do not shuffle content across files without a payoff.
 
 ### Change events, research, or partners data shape
 
-Start in `packages/notion-data/src/index.ts`.
+Start in `src/sanity/schemas/`.
 
 Typical flow:
 
-1. update the mapper or type
-2. update `src/lib/notion.ts` consumers if needed
-3. update views/components that render the changed fields
-4. update `test/notion-data.test.ts`
-5. run `pnpm test` and `pnpm typecheck:all`
+1. update the schema definition (e.g., `src/sanity/schemas/event.ts`)
+2. update the GROQ queries in `src/lib/sanity.ts`
+3. update the TypeScript interface in `src/lib/types.ts`
+4. update views/components that render the changed fields
+5. run `pnpm typecheck`
 
 Important:
 
-- `/events` and `/research` render from server-side calls to `src/lib/notion.ts`
+- `/events`, `/research`, and `/partners` render from server-side calls to `src/lib/sanity.ts`
 - the API routes mirror that data, but they are not the primary rendering path
 
 ### Add or change a JSON endpoint
@@ -87,7 +86,7 @@ Keep them thin:
 
 - read shared data from `src/lib/`
 - return `NextResponse.json(...)`
-- do not duplicate Notion querying logic inside the route handler
+- do not duplicate Sanity querying logic inside the route handler
 
 ### Change styling or theming
 
@@ -109,7 +108,7 @@ Put shipped files under `public/assets/` in the closest existing subfolder.
 Examples:
 
 - route-specific photos in `public/assets/homepage/`, `public/assets/apply/`, `public/assets/e-lab/`
-- logos in partner- or sponsor-specific folders
+- static logos in `public/assets/partners_sponsors/` (note: dynamic partners are handled in Sanity)
 
 Use brand source files in `docs/brand/source/` as reference material, not as runtime assets.
 
@@ -127,7 +126,7 @@ For a new page, update both:
 - `src/data/routes.tsx` is legacy and not used for runtime routing
 - `/api/getNotes` returns events despite the name
 - `src/app/globals.css` only re-imports `src/styles/index.css`
-- changing only an API route will not affect a page that fetches directly from `src/lib/notion.ts`
+- changing only an API route will not affect a page that fetches directly from `src/lib/sanity.ts`
 
 ## Suggested Review Checklist
 
@@ -136,6 +135,5 @@ Before shipping a change, check:
 - correct ownership layer touched
 - no duplicated content or data-fetching logic introduced
 - `src/config/seo.ts` updated if a route changed
-- tests updated if types or mapping changed
 - `pnpm verify` passed
-- `pnpm typecheck:all` passed when TypeScript contracts changed
+- `pnpm typecheck` passed when TypeScript contracts changed
