--- packages/notion-data/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@tumai/notion-data",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "exports": {
-    ".": "./src/index.ts"
-  },
-  "scripts": {
-    "lint": "biome check src",
-    "typecheck": "tsc --noEmit -p tsconfig.json"
-  },
-  "dependencies": {
-    "@notionhq/client": "^3.1.3"
-  },
-  "devDependencies": {
-    "@types/node": "^22.15.29",
-    "typescript": "~5.8.3"
-  }
-}
