--- pnpm-workspace.yaml
@@ -1,3 +1,0 @@
-packages:
-  - "."
-  - "packages/*"
